suspects.py changed to correct_answer.py and refactored
