+++ main.py
@@ -9,6 +9,7 @@ from src.screens.game_won import GameWon
 from src.screens.main_menu import MainMenu
 from src.screens.rules import Rules
 from src.screens.questions import Question
+from src.screens.correct_answer import CorrectAnswer
 from src.screens.suspects import Suspects
 from src.timer import Timer
 from src.screens.wrong_answer import WrongAnswer
@@ -32,13 +33,12 @@ class Game:
         self.main_menu = MainMenu(self.screen, self.game_state_manager, self.draw)
         self.rules = Rules(self.screen, self.game_state_manager, self.draw, self.timer)
         self.question = Question(self.screen, self.game_state_manager, self.draw, self.timer)
+        self.suspects = Suspects(self.screen, self.game_state_manager, self.draw, self.murderer)
         self.wrong_answer = WrongAnswer(self.screen, self.game_state_manager, self.draw, self.timer, self.question,
                                         self.button_handler)
-        self.suspects = Suspects(self.screen, self.game_state_manager, self.draw, self.timer, self.question, self.clues,
-                                 self.murderer,
-                                 self.button_handler)
-        self.arrest_suspect = ArrestSuspect(self.screen, self.game_state_manager, self.draw, self.timer,
-                                            self.button_handler, self.suspects)
+        self.correct_answer = CorrectAnswer(self.screen, self.game_state_manager, self.draw, self.timer, self.question,
+                                            self.clues, self.button_handler, self.suspects)
+        self.arrest_suspect = ArrestSuspect(self.screen, self.game_state_manager, self.draw, self.timer, self.suspects)
         self.game_lost = GameLost(self.screen, self.game_state_manager, self.draw)
         self.game_won = GameWon(self.screen, self.game_state_manager, self.draw)
 
@@ -47,7 +47,7 @@ class Game:
             "rules": self.rules,
             "question": self.question,
             "wrong_answer": self.wrong_answer,
-            "suspects": self.suspects,
+            "suspects": self.correct_answer,
             "arrest_suspect": self.arrest_suspect,
             "game_won": self.game_won,
             "game_lost": self.game_lost
